Move helpers to the folder where they are used
