+++ packages/shopify-app-remix/src/server/auth/admin/authenticate.ts
@@ -18,31 +18,33 @@ import type {
   AppConfig,
   AppConfigArg,
 } from '../../config-types';
-import type {BillingContext} from '../../billing/types';
 import {
-  cancelBillingFactory,
-  requestBillingFactory,
-  requireBillingFactory,
-} from '../../billing';
-import {
-  beginAuth,
   getSessionTokenHeader,
-  redirectWithExitIframe,
-  redirectToAuthPage,
   validateSessionToken,
   rejectBotRequest,
   respondToOptionsRequest,
   ensureCORSHeadersFactory,
-  renderAppBridge,
-  redirectFactory,
-  handleClientErrorFactory,
 } from '../helpers';
 
+import type {BillingContext} from './billing/types';
+import {
+  cancelBillingFactory,
+  requestBillingFactory,
+  requireBillingFactory,
+} from './billing';
 import type {
   AdminContext,
   EmbeddedAdminContext,
   NonEmbeddedAdminContext,
 } from './types';
+import {
+  beginAuth,
+  handleClientErrorFactory,
+  redirectFactory,
+  redirectToAuthPage,
+  redirectWithExitIframe,
+  renderAppBridge,
+} from './helpers';
 
 interface SessionContext {
   session: Session;
